requirements.txt
# Minimal requirements for AI Research Agent

# Core dependencies
langchain-core
langchain-community
langchain-openai

# LangGraph with multiserver support
langgraph>=0.1.1
langchain-mcp-adapters>=0.0.1

# API and deployment
fastapi
uvicorn

# Environment and utilities
python-dotenv
requests

# Research tools
tavily-python

